fix(msvs): visual studio project files remove win32 and add fpp
The Visual Studio Solution (.sln) and Project (.vfproj) files
no longer include the win32 compilation option. The files
are also updated to Visual Studio 2022 format. This format
still works with an up to date Visual Studio 2019.

The mf5to6 and zonebudget projects added a missing
fpp option for successful release compilation. Intermediate
files are now placed in an obj_temp directory and the
resulting executables are placed in the bin folder
(previously, a "copy" post-build command was invoked).

The debug symbols are only generated for debug compilation
and now are placed in the obj_temp directory. Release
builds only include line numbers for stack tracing.